mod03 aula42

+++ sistema/painel-admin/paginas/cursos/excluir.php
@@ -8,7 +8,13 @@ $id = $_POST['id'];
 //para apagar foto caso tenha sido feito upload dela
 $query = $pdo->query("SELECT * FROM $tabela WHERE id='$id'");
 $res = $query->fetchAll(PDO::FETCH_ASSOC);
-$foto = $res[0]['foto'];
+$foto = $res[0]['imagem'];
+$status = $res[0]['status'];
+
+if($status == 'Aprovado') {
+    echo 'Esse curso foi APROVADO e por isso não pode ser excluído. Para excluí-lo, altere o status dele para AGUARDANDO';
+    exit();
+}
 
 if($foto != 'sem-foto.png') {
     unlink('../../img/cursos/'.$foto);
